use common vectorlayerpresentation component
