+++ bundles/admin/admin-layereditor/view/AdminLayerForm/AdditionalTabPane/VectorLayerAttributes.jsx
@@ -4,12 +4,12 @@ import styled from 'styled-components';
 import { Message, Select, Button, Badge } from 'oskari-ui';
 import { Modal } from 'oskari-ui/components/Modal';
 import { FeatureFilter, cleanFilter } from 'oskari-ui/components/FeatureFilter';
+import { VectorLayerPresentation } from 'oskari-ui/components/VectorLayerPresentation';
 import { InfoIcon } from 'oskari-ui/components/icons';
 import { Controller, Messaging } from 'oskari-ui/util';
 import { PropertiesFilter, PropertiesLocale, PropertiesFormat } from './VectorLayerAttributes/';
 import { StyledFormField, Border } from '../styled';
 import { GEOMETRY_TYPES, DATA, getGeometryType } from '../../LayerHelper';
-
 const Buttons = styled.div`
     display: inline-flex;
     > * {
@@ -38,26 +38,7 @@ const clean = obj => {
 };
 
 export const VectorLayerAttributes = ({ layer, controller }) => {
-    const { data = {}, filter: featureFilter } = layer.attributes;
-    const { geomName, featureProperties = []} = layer.capabilities;
-    const [modal, setModal] = useState(null);
-    const [state, setState] = useState({
-        filter: data.filter || {},
-        locale: data.locale || {},
-        format: data.format || {},
-        featureFilter
-    });
-    const getButtonForModal = type => {
-        const value = state[type] || {};
-        const count = Object.keys(value).length;
-        return (
-            <Badge count={count} showZero={false}>
-                <Button onClick={() => onButtonClick(type)}>
-                    <Message messageKey={`attributes.${type}.button`} />
-                </Button>
-            </Badge>
-        );
-    };
+    const { data = {} } = layer.attributes;
 
     const onGeometryTypeChange = value => {
         const key = DATA.GEOMETRY;
@@ -67,50 +48,16 @@ export const VectorLayerAttributes = ({ layer, controller }) => {
             controller.setAttributesData(key, value);
         }
     };
-    const onModalOk = () => {
-        if (modal === 'featureFilter') {
-            const filter = cleanFilter(state.featureFilter, featureProperties);
-            controller.setFeatureFilter(filter);
-            // update local state
-            onModalUpdate(filter);
-            setModal(null);
-            return;
-        }
-        // deep clone to not mess local state
-        const value = JSON.parse(JSON.stringify(state[modal]));
-        // clean twice to get rid of empty objects if last value is deleted from it
-        clean(value);
-        clean(value);
-        if (Object.keys(value).length) {
-            controller.setAttributesData(modal, value);
-        } else {
-            controller.setAttributesData(modal);
-        }
-        // update local state
-        onModalUpdate(value);
-        setModal(null);
+
+    const updateAttributes = (modal, value) => {
+        controller.setAttributesData(modal,value);
     };
-    const onButtonClick = mode => {
-        if (!featureProperties.length) {
-            Messaging.warn(<Message messageKey='messages.noFeatureProperties' bundleKey='admin-layereditor'/>);
-            return;
-        }
-        setModal(mode);
+
+    const updateFeatureFilter = (value) => {
+        controller.setFeatureFilter(value);
     };
-    const onModalUpdate = (value) => {
-        setState({ ...state, [modal]: value });
-    };
-    const onModalCancel = () => {
-        const attr = modal === 'featureFilter' ? featureFilter : data[modal] || {};
-        setState({ ...state, [modal]: attr });
-        setModal(null);
-    };
-    const properties = featureProperties.filter(prop => prop.name !== geomName);
-    const propNames = properties.map(prop => prop.name);
+
     const geometryTypeSource = data[DATA.GEOMETRY] ? 'Attributes' : 'Capabilities';
-    const propLabels = Oskari.getLocalized(data.locale) || {};
-    // gather selected properties from all (localized) filters
-    const selectedProperties = state.filter ? [...new Set([].concat(...Object.values(state.filter)))] : [];
     return (
         <Fragment>
             <Message messageKey='attributes.geometryType.label'/>
@@ -125,56 +72,8 @@ export const VectorLayerAttributes = ({ layer, controller }) => {
                     }))}
                 />
             </StyledFormField>
-            <Message messageKey='attributes.properties'/>
-            <Border>
-                <StyledFormField>
-                    { getButtonForModal('featureFilter') }
-                </StyledFormField>
-                <Message messageKey='attributes.presentation' />
-                <InfoIcon title={<Message messageKey='attributes.presentationTooltip'/>}/>
-                <StyledFormField>
-                    <Buttons>
-                        { getButtonForModal('filter') }
-                        { getButtonForModal('locale') }
-                        { getButtonForModal('format') }
-                    </Buttons>
-                </StyledFormField>
-                <Message messageKey='attributes.idProperty'/>
-                <InfoIcon title={<Message messageKey='attributes.idPropertyTooltip'/>}/>
-                <StyledFormField>
-                    <Select allowClear value={data[DATA.REPLACE_ID]}
-                        onChange={value => controller.setAttributesData(DATA.REPLACE_ID, value)}
-                        options={propNames.map(value => ({value}))}/>
-                </StyledFormField>
-            </Border>
-            <Modal
-                mask={ false }
-                maskClosable= { false }
-                open={ !!modal }
-                onOk={ onModalOk}
-                onCancel={ onModalCancel }
-                cancelText={ <Message messageKey="cancel" /> }
-                okText={ <Message messageKey="save" /> }
-                width={ modal === 'featureFilter' ? 800 : 500 }
-            >
-                <h3><Message messageKey={`attributes.${modal}.title`} /></h3>
-                { modal === 'filter' &&
-                    <PropertiesFilter update={onModalUpdate} properties={propNames}
-                        filter={state.filter} labels={propLabels}/>
-                }
-                { modal === 'locale' &&
-                    <PropertiesLocale update={onModalUpdate} locale={state.locale}
-                        properties={propNames} selected={selectedProperties}/>
-                }
-                { modal === 'format' &&
-                    <PropertiesFormat update={onModalUpdate} properties={propNames}
-                        format={state.format} labels={propLabels} selected={selectedProperties}/>
-                }
-                { modal === 'featureFilter' &&
-                    <FeatureFilter onChange={onModalUpdate} properties={propNames}
-                        filter={state.featureFilter} labels={propLabels} types={properties}/>
-                }
-            </Modal>
+
+            <VectorLayerPresentation layer={layer} updateAttributes={updateAttributes} updateFeatureFilter={updateFeatureFilter} />
         </Fragment>
     );
 };
